Refactoring of ScriptRegistry related classes
