+++ modules/mono/glue/GodotSharp/GodotSharp/Core/Bridge/CSharpInstanceBridge.cs
@@ -49,7 +49,6 @@ namespace Godot.Bridge
                     retValue = scriptMethod(godotObject, in argsStruct);
 
                     return true;
-
                 }
                 else
                 {
