+++ modules/mono/glue/GodotSharp/GodotSharp/Core/Bridge/ScriptPropertyMethod.cs
@@ -1,0 +1,7 @@
+using Godot.NativeInterop;
+
+namespace Godot.Bridge
+{
+    public delegate godot_variant ScriptPropertyMethod<T>(T godotObject, scoped in godot_variant value)
+        where T : GodotObject;
+}
