+++ modules/mono/glue/GodotSharp/GodotSharp/Core/Bridge/ScriptPropertyRegistry.cs
@@ -1,0 +1,13 @@
+namespace Godot.Bridge
+{
+    public sealed class ScriptPropertyRegistry<T> :
+        ScriptRegistry<T, ScriptPropertyMethod<GodotObject>, ScriptCache<T, ScriptPropertyMethod<GodotObject>>,
+            ScriptPropertyRegistry<T>>
+        where T : GodotObject
+    {
+        protected override void InitializeCache((MethodKey, ScriptPropertyMethod<GodotObject>)[] methods)
+        {
+            ScriptCache<T, ScriptPropertyMethod<GodotObject>>.Initialize(methods, []);
+        }
+    }
+}
