+++ modules/mono/glue/GodotSharp/GodotSharp/Core/Bridge/ScriptSignalMethod.cs
@@ -1,0 +1,7 @@
+using Godot.NativeInterop;
+
+namespace Godot.Bridge
+{
+    public delegate void ScriptSignalMethod<T>(T godotObject, scoped in NativeVariantPtrArgs args)
+        where T : GodotObject;
+}
